services/ingest-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
aiofiles==23.2.1
PyMuPDF==1.23.8
openai==1.3.7
weaviate-client==3.25.3
python-docx==1.1.0
python-pptx==0.6.23
openai-whisper==20231117
tiktoken==0.5.2
prometheus-client==0.19.0
python-magic==0.4.27
email-validator==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4